requirements.txt
numpy
rasterio
scikit-learn
matplotlib
seaborn
pandas
geopandas
pyyaml
shapely
folium
branca